Added Foul Soup
Foul Soup is crafted with two Foul Berries, a Spider Eye and a Bowl. It gives the player two hunger points and Foul Taste II when eaten. It should now be possible to have both Foul Taste and Vanilla Scent and the same time, making the "How Did We Get Here?" advancement possible.